refactor/parser-strategies разделено по каталогам
