+++ src/core/ParserStrategy.js
