+++ src/strategies/index.js
@@ -1,0 +1,3 @@
+export { AuchanStrategy } from './AuchanStrategy';
+export { DeliveryClubStrategy } from './DeliveryClubStrategy';
+export { MagnitStrategy } from './MagnitStrategy';
